Silence "cp: cannot stat '/etc/cmdline'..." error message
These files may or may not exist; check to see if they exist before
copying them!

+++ dracut/anaconda-copy-cmdline.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 # Copy over cmdline(.d) files from the initrd to /run before pivot
 mkdir -p /run/install/cmdline.d
-cp /etc/cmdline.d/* /run/install/cmdline.d/
-cp /etc/cmdline /run/install/
+for f in /etc/cmdline.d/*; do
+    [ -e $f ] && cp $f /run/install/cmdline.d/
+done
+[ -e /etc/cmdline ] && cp /etc/cmdline /run/install/
